Test for Matrix moved.
