+++ src/se/hh/aoop/assignment1/exercise4/Stack.java
@@ -24,7 +24,9 @@ public class Stack {
   }
 
   public int pop(int n) {
-    return this.stack
+    return 0;
   }
+
+
 
 }
